requirements.txt
llama-index
llama-index-core
llama-index-embeddings-openai
pinecone-client
llama-index-vector-stores-pinecone
python-dotenv
langchain
flask
boto3
urllib3==1.26.7
langchain 
langchain-openai
requests 
pandas 
beautifulsoup4
networkx
